+++ modules/indexing/src/test/java/org/apache/ignite/testsuites/IgnitePdsWithIndexingTestSuite.java
@@ -41,7 +41,7 @@ import org.apache.ignite.internal.processors.database.IgniteTwoRegionsRebuildInd
 import org.apache.ignite.internal.processors.database.RebuildIndexTest;
 import org.apache.ignite.internal.processors.database.RebuildIndexWithHistoricalRebalanceTest;
 import org.apache.ignite.internal.processors.database.RebuildIndexWithMVCCTest;
-import org.apache.ignite.internal.processors.query.h2.maintenance.MaintenanceRebuildIndexTargetSelfTest;
+import org.apache.ignite.internal.processors.query.h2.maintenance.MaintenanceRebuildIndexUtilsSelfTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
@@ -73,7 +73,7 @@ import org.junit.runners.Suite;
     RenameIndexTreeTest.class,
     DropIndexTest.class,
     AffinityIndexShadowingTest.class,
-    MaintenanceRebuildIndexTargetSelfTest.class
+    MaintenanceRebuildIndexUtilsSelfTest.class
 })
 public class IgnitePdsWithIndexingTestSuite {
 }
